% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2f.R
\name{c2f}
\alias{c2f}
\title{Convert temperature in Fahrenheit into Celsius}
\usage{
c2f(c)
}
\arguments{
\item{c}{Temperature in C}
}
\value{
The temperature in F
}
\description{
Convert temperature in Fahrenheit into Celsius
}
\examples{
c2f(100)
}
