% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello2.R
\name{hello2}
\alias{hello2}
\title{A simple function doing little}
\usage{
hello2(txt = "world")
}
\arguments{
\item{txt}{A text argument}
}
\value{
Nothing, side effect of cat()
}
\description{
A simple function doing little
}
\examples{
hello("world")
}
