\name{sampleTestPackage-package}
\alias{sampleTestPackage-package}
\alias{sampleTestPackage}
\docType{package}
\title{\packageTitle{sampleTestPackage}}
\description{\packageDescription{sampleTestPackage}}
\details{
  The DESCRIPTION file: \packageDESCRIPTION{sampleTestPackage}
  \packageIndices{sampleTestPackage}

  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
  \packageAuthor{sampleTestPackage}
  
  Maintainer: \packageMaintainer{sampleTestPackage}
}
\keyword{package}
