\name{sampleTestPackage-package}
\alias{sampleTestPackage-package}
\alias{sampleTestPackage}
\docType{package}
\title{\packageTitle{sampleTestPackage}}
\description{\packageDescription{sampleTestPackage}}
\details{
  The DESCRIPTION file: \packageDESCRIPTION{sampleTestPackage}
  \packageIndices{sampleTestPackage}

  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
  \packageAuthor{sampleTestPackage}
  
  Maintainer: \packageMaintainer{sampleTestPackage}
}
\references{
  This optional section can contain literature or other references for
  background information.
}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{package}
\seealso{
  Optional links to other man pages
}
\examples{
  ## Optional simple examples of the most important functions
  ## Use \dontrun{} around code to be shown but not executed
}
